fix 2 issues
1) readtmsCheckGC was incorrectly asking readtms to ignore GC.
The intent was to pass back as a result whether the call GC'd or not,
so the caller could choose how to handle it.
It led to dafilesrv not closing the socket after a GC in this situation,
causing a spurious follow-on error on a notifySelect.

2) UDP reading should not have tried to be reading more data if it
had already received something. This was not a new issue in itself,
but the semantics changed to keep reading beyond min_size until blocked.
It makes no sense to read if have already received >0 from recvfrom when
dealing with UDP. Altered the logic and added comments re. semantics.